--- starter-code/e2e/app.e2e-spec.ts
@@ -1,14 +1,0 @@
-import { ContactAppPage } from './app.po';
-
-describe('contact-app App', function() {
-  let page: ContactAppPage;
-
-  beforeEach(() => {
-    page = new ContactAppPage();
-  });
-
-  it('should display message saying app works', () => {
-    page.navigateTo();
-    expect(page.getParagraphText()).toEqual('app works!');
-  });
-});
