--- starter-code/src/app/app.component.html
@@ -1,4 +1,0 @@
-<h1>
-  {{title}}
-</h1>
-<app-contact-list></app-contact-list>
