+++ starterCode/src/app/app.component.html
@@ -1,0 +1,2 @@
+
+<app-contact-list></app-contact-list>
