+++ starterCode/src/app/app.module.ts
@@ -1,0 +1,24 @@
+import { BrowserModule } from '@angular/platform-browser';
+import { NgModule } from '@angular/core';
+import { FormsModule } from '@angular/forms';
+import { HttpModule } from '@angular/http';
+
+
+import { AppComponent } from './app.component';
+import { ContactListComponent } from './contact-list/contact-list.component';
+
+
+@NgModule({
+  declarations: [
+    AppComponent,
+    ContactListComponent
+  ],
+  imports: [
+    BrowserModule,
+    FormsModule,
+    HttpModule
+  ],
+  providers: [],
+  bootstrap: [AppComponent]
+})
+export class AppModule { }
