+++ starterCode/src/app/contact-list/contact-list.component.css
@@ -1,0 +1,16 @@
+img {
+  width:100px;
+}
+
+.container {
+  display:flex;
+  align-items:center;
+  flex-direction:column;
+  margin:0;
+  
+}
+
+.border {
+  border:1px border solid;
+}
+
